requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.2
cachetools==4.2.0
certifi==2020.12.5
cfbclient==1.1.0
cffi==1.14.4
chardet==4.0.0
click==8.0.4
cryptography==2.1.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
diplomat==1.4.0
entrypoints==0.3
faiss==1.5.3
filelock==3.4.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gr-record==1.12
grpcio==1.32.0
gviz-api==1.9.0
h5py==2.10.0
huggingface-hub==0.0.12
idna==2.6
importlib-metadata==3.4.0
importlib-resources==5.4.0
ipykernel==5.1.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
lightgbm==3.2.1
llvmlite==0.36.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
missingno==0.5.0
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==4.4.0
nest-asyncio==1.4.3
nltk==3.6.7
notebook==6.2.0
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==5.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.13
protobuf==3.14.0
pstatsd==1.2.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.4
pygobject==3.26.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-consul==1.1.0
python-dateutil==2.8.1
pytz==2021.1
pyxdg==0.25
PyYAML==6.0
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
redis==3.5.3
regex==2022.3.15
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.49
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
SecretStorage==2.3.1
Send2Trash==1.5.0
six==1.15.0
skipthoughts==0.0.1
ssh-import-id==5.7
tenacity==8.0.1
tensorboard==2.4.1
tensorboard-logger==0.1.0
tensorboard-plugin-profile==2.5.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.2.0
timm==0.5.4
tlcache==0.3.1
tokenizers==0.10.3
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.63.1
traitlets==4.3.3
transformers==4.8.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
zipp==3.4.0
